Move total yearly EV energy usage to transportation calculations file and reducer
